# [stdlib] Optimize `.splitlines()` using new `splat` bit mask util