--- a/mojo/stdlib/stdlib/collections/string/codepoint.mojo
+++ b/mojo/stdlib/stdlib/collections/string/codepoint.mojo
@@ -44,6 +44,7 @@ var s = String(c)  # "A"
 from sys.intrinsics import likely
 
 from bit import count_leading_zeros
+from bit._mask import splat
 
 
 @always_inline
@@ -561,7 +562,7 @@ struct Codepoint(Copyable, EqualityComparable, Intable, Movable, Stringable):
                 var mask = UInt8(0xFF) >> (num_bytes + Int(num_bytes > 1))
                 var num_bytes_marker = UInt8(0xFF) << (8 - num_bytes)
                 ptr[0] = ((c >> shift) & mask) | (
-                    num_bytes_marker & -Int(num_bytes != 1)
+                    num_bytes_marker & splat(num_bytes != 1)
                 )
                 for i in range(1, num_bytes):
                     shift -= 6