Optimize `.splitlines()` using new `splat` bit mask util


### Benchmark results

CPU: Intel® Core™ i7-7700HQ

|  | old value (ms) | new value (ms) | percentage | magnitude
|-- | --:| --:| --:| --:|
|bench_string_splitlines[10] | 9.93 | 8.76 | 0.12 | 1.13|
|bench_string_splitlines[30] | 6.46 | 5.00 | 0.23 | 1.29|
|bench_string_splitlines[50] | 4.91 | 4.60 | 0.06 | 1.07|
|bench_string_splitlines[100] | 4.49 | 3.85 | 0.14 | 1.17|
|bench_string_splitlines[1000] | 3.98 | 3.57 | 0.10 | 1.12|
|bench_string_splitlines[10000] | 3.67 | 3.36 | 0.08 | 1.09|
|bench_string_splitlines[100000] | 3.67 | 3.34 | 0.09 | 1.10|
|bench_string_splitlines[1000000] | 3.67 | 3.42 | 0.07 | 1.07|
|average improvement |   |   | 0.11 | 1.13|