@laszlokindrat 2025-08-26
!sync